refactor : service layer interface impl
